+++ Rdmp.Core/CommandExecution/AtomicCommands/ExecuteCommandDelete.cs
@@ -20,20 +20,29 @@ namespace Rdmp.Core.CommandExecution.AtomicCommands
     public class ExecuteCommandDelete : BasicCommandExecution
     {
         private readonly IList<IDeleteable> _deletables;
+        
+        /// <summary>
+        /// Flag applies only for deletion where the UI layer is non-interactive.  True to allow
+        /// multiple deletes to go ahead without asking.  False to throw exception
+        /// </summary>
+        private readonly bool _allowDeleteMany;
 
-        [UseWithObjectConstructor]
         public ExecuteCommandDelete(IBasicActivateItems activator, 
-            [DemandsInitialization("The object you want to delete",Mandatory = true)]
             IDeleteable deletable) : this(activator,new []{ deletable})
         {
         }
 
-        
-        public ExecuteCommandDelete(IBasicActivateItems activator, IDeleteable[] deletables) : base(activator)
+
+        [UseWithObjectConstructor]
+        public ExecuteCommandDelete(IBasicActivateItems activator,
+            [DemandsInitialization("The object you want to delete",Mandatory = true)]
+            IDeleteable[] deletables,
+            [DemandsInitialization("Optional.  Pass \"true\" to allow deleting many objects at once e.g. Catalogue:*bob* (deletes all catalogues with the word bob in)")]
+            bool deleteMany = false) : base(activator)
         {
             _deletables = deletables;
-
-            if(_deletables.Any( d => d is CohortAggregateContainer c && c.IsRootContainer()))
+            this._allowDeleteMany = deleteMany;
+            if (_deletables.Any( d => d is CohortAggregateContainer c && c.IsRootContainer()))
                 SetImpossible("Cannot delete root containers");
             
             string reason = "";
@@ -50,7 +59,15 @@ namespace Rdmp.Core.CommandExecution.AtomicCommands
             else
             if(_deletables.Count>1)
             {
-                if(YesNo("Delete " + _deletables.Count + " Items?","Delete Items"))
+                // if the command did not ask to delete many and it is not interactive (e.g. CLI) then 
+                // we shouldn't just blindly delete them all
+                if (!BasicActivator.IsInteractive && !_allowDeleteMany)
+                {
+                    throw new Exception($"Allow delete many is false but mutliple objects were matched for deletion ({string.Join(",",_deletables)})");
+                }
+
+                // if it is not interactive or user confirms behaviour
+                if(!BasicActivator.IsInteractive || YesNo("Delete " + _deletables.Count + " Items?","Delete Items"))
                 {
                     try
                     {