+++ CHANGELOG.md
@@ -19,13 +19,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - IPluginUserInterface can no longer add items to tab menu bars (only context menus)
 - Removed some Catalogue context menu items when the Catalogue is an API call
 - Adding a Filter from Catalogue no longer opens it up in edit mode after adding
+- Command line execution (e.g. `rdmp cmd ...`) no longer supports user interactive calls (e.g. YesNo questions)
 
 ### Added
 
 - Added custom metadata report tokens:
   - $Comma (for use with formats that require seperation e.g. JSON when using the `$foreach` operation)
   - $TimeCoverage_ExtractionInformation (the column that provides the time element of a dataset to the DQE e.g. StudyDate)
-
+- Added support for default values in constructors invoked from the command line (previously command line had to specify all arguments.  Now you can skip default ones at the end of the line)
+- Added support for deleting mutliple objects at once with the delete command (e.g. `rdmp cmd Delete Plugin true` to delete all plugins)
+  - Boolean flag at the end is optional and defaults to false (expect to delete only 1 object)
+  - Use `rdmp cmd DescribeCommand Delete` for more information
 
 ## [6.0.2] - 2021-08-26